Issue #812 Non-functional library manager refactor
- Replace static methods with free functions in anonymous namespace
- Pull implementations out of the header file in to a new .cpp file

No logical changes.
